Revert "(PC-3976): Refacto clean architecture"
This reverts commit f86470e966830bb1010eec10f471b5df621d0ec9.
